QA FORM COMPLETED WITH FORMATTING
This is the user submit form, I formatted the textarea completely, removed its borders and background, added a bottom border to match the color I think goes well with the site. The div container is set to a 32.5% to give it an offset from the pre existing sidebar menu. 25% will center it, but I thought it was better to display like this. 
Also, The background image is placed in the CSS html{}, this is only so you can see the design before implimented. Everything is transparent so it can just be dropped into the existing design. 

Thanks again!